Clean up domain transaction pool
